Tidy cmdscale() list output

These tidiers handle the output of cmdscale(), which under certain conditions is effectively an S3 object without a class attribute. It allows ordr to enhance the list_tidiers provided by broom.

tidy_cmdscale(x, matrix = "points", ...)

glance_cmdscale(x, ...)

Arguments

x

A list with components points, eig, x, ac, and GOF returned by stats::cmdscale().

matrix

Character specifying which list element should be tidied, matched to the following options.

"points": returns information about the coordinates in the representation space.

"x": returns information about the doubly-centered symmetric matrix used in the calculation.

"eig": returns information about the eigenvalues.

...

Additional arguments allowed by generics; currently ignored.

See also

Examples

# 'dist' object (matrix of road distances) of large American cities
class(UScitiesD)
#> [1] "dist"
print(UScitiesD)
#>               Atlanta Chicago Denver Houston LosAngeles Miami NewYork
#> Chicago           587                                                
#> Denver           1212     920                                        
#> Houston           701     940    879                                 
#> LosAngeles       1936    1745    831    1374                         
#> Miami             604    1188   1726     968       2339              
#> NewYork           748     713   1631    1420       2451  1092        
#> SanFrancisco     2139    1858    949    1645        347  2594    2571
#> Seattle          2182    1737   1021    1891        959  2734    2408
#> Washington.DC     543     597   1494    1220       2300   923     205
#>               SanFrancisco Seattle
#> Chicago                           
#> Denver                            
#> Houston                           
#> LosAngeles                        
#> Miami                             
#> NewYork                           
#> SanFrancisco                      
#> Seattle                678        
#> Washington.DC         2442    2329
# use multidimensional scaling to infer artificial planar coordinates
UScitiesD %>%
  cmdscale(k = 3L, eig = TRUE, x.ret = TRUE) ->
  usa_mds
# return the tidied coordinates
tidy(usa_mds)
#> Error: No tidy method recognized for this list.
# return the upper triangle of the doubly-centered distance matrix in tidy form
tidy(usa_mds, matrix = "x")
#> Error: No tidy method recognized for this list.
# return the eigenvalues for the principal coordinates, with summary statistics
tidy(usa_mds, matrix = "eig")
#> Error: No tidy method recognized for this list.
# scree plot of inertia
ggplot(tidy(usa_mds, matrix = "e"), aes(x = PCo, y = eig)) +
  theme_bw() +
  geom_col() +
  labs(x = "Principal coordinate", y = "Eigenvalue (inertia)")
#> Error: No tidy method recognized for this list.
# reorient to conventional compass
tidy(usa_mds) %>%
  ggplot(aes(x = -PCo1, y = -PCo2)) +
  geom_text(aes(label = point), size = 3) +
  ggtitle("MDS biplot of distances between U.S. cities")
#> Error: No tidy method recognized for this list.