Tidying methods for (additional) lists

Some unclassed lists of fixed structure, returned by other functions, are recognized by broom's list tidiers. ordr's updated list tidiers append additional checks for the list structures returned by cmdscale() and cancor(). If such structure is detected, then the new in-package tidiers are called.

# S3 method for list
tidy(x, ...)

# S3 method for list
glance(x, ...)

Arguments

x

A list, potentially representing an object that can be tidied.

...

Additionally, arguments passed to the tidying function.

Value

A tibble.

See also