Masks of external ordination objects

Some popular ordination functions in R return objects without a specific class (e.g. stats::cmdscale()). Some others return output that is not self-contained, for instance missing annotation from the original dataset (e.g. logisticPCA::logisticPCA()). The functions below wrap the originals with the minimal overhead necessary for complete functionality with the "[tbl_ord]" class.

cmdscale(d, k = 2, add = FALSE)

logisticSVD(x, k = 2, quiet = TRUE, max_iters = 1000,
  conv_criteria = 1e-05, random_start = FALSE, ...,
  partial_decomp = TRUE, main_effects = TRUE)

logisticPCA(x, k = 2, m = 4, quiet = TRUE, partial_decomp = FALSE,
  max_iters = 1000, conv_criteria = 1e-05, random_start = FALSE, ...,
  main_effects = TRUE)

convexLogisticPCA(x, k = 2, m = 4, quiet = TRUE,
  partial_decomp = FALSE, max_iters = 1000, conv_criteria = 1e-06,
  random_start = FALSE, ..., main_effects = TRUE, ss_factor = 4)

Arguments

d
a distance structure such as that returned by dist or a full symmetric matrix containing the dissimilarities.

k
the maximum dimension of the space which the data are to be represented in; must be in \(\{1, 2, \ldots, n-1\}\).

add
logical indicating if an additive constant \(c*\) should be computed, and added to the non-diagonal dissimilarities such that the modified dissimilarities are Euclidean.

x
matrix with all binary entries

quiet
logical; whether the calculation should give feedback

max_iters
number of maximum iterations

conv_criteria
convergence criteria. The difference between average deviance in successive iterations

random_start
logical; whether to randomly inititalize the parameters. If FALSE, function will use an eigen-decomposition as starting value

...
Additional parameters passed to original functions.

partial_decomp
logical; if TRUE, the function uses the rARPACK package to more quickly calculate the eigen-decomposition. This is usually faster than standard eigen-decomponsition when ncol(x) > 100 and k is small

main_effects
logical; whether to include main effects in the model

m
value to approximate the saturated model. If m = 0, m is solved for

ss_factor
step size multiplier. Amount by which to multiply the step size. Quadratic convergence rate can be proven for ss_factor = 1, but I have found higher values sometimes work better. The default is ss_factor = 4. If it is not converging, try ss_factor = 1.