Tidiers for 'tbl_ord' objects

These functions return tibbles that summarize an object of class 'tbl_ord'. tidy() output contains one row per artificial coordinate and glance() output contains one row for the whole ordination.

# S3 method for tbl_ord
tidy(x, ...)

# S3 method for tbl_ord
glance(x, ...)

# S3 method for tbl_ord
fortify(model, data, ..., .matrix = "dims", .supplement = TRUE)

Arguments

x, model

An object of class 'tbl_ord'.

...

Additional arguments allowed by generics; currently ignored.

data

Passed to generic methods; currently ignored.

.matrix

A character string matching one of several indicators for one or both matrices in a matrix decomposition used for ordination. The standard values are "rows", "cols", and "dims" (for both).

.supplement

Logical; whether to include supplementary points.

Details

Three generics popularized by the ggplot2 and broom packages make use of the augmentation methods:

The generics::tidy() method summarizes information about model components, which here are the artificial coordinates created by ordinations. The output can be passed to ggplot2::ggplot() to generate scree plots.

The generics::glance() method reports information about the entire model, here always treated as one of a broader class of ordination models.

The ggplot2::fortify() method augments and collapses row and/or column data, depending on .matrix, into a single tibble, in preparation for ggplot2::ggplot(). Its output resembles that of generics::augment(), though rows in the output may correspond to rows, columns, or both of the original data. If .matrix is passed "rows", "cols", or "dims" (for both), then fortify() returns a tibble whose fields are obtained, in order, via get_*(), augmentation_*(), and annotation_*().

The tibble is assigned a "coordinates" attribute whose value is obtained via get_coord(). This facilitates some downstream functionality that relies on more than those coordinates used as position aesthetics in a biplot, in particular stat_spantree().

See also