Confer inertia to factors of a 'tbl_ord' object

Re-distribute inertia between cases and variables in an ordination.

recover_conference(x)

# S3 method for default
recover_conference(x)

get_conference(x)

revert_conference(x)

confer_inertia(x, p)

Arguments

x

A tbl_ord object.

p

Numeric vector of length 1 or 2. If length 1, the proportion of the inertia assigned to the cases, with the remainder 1 - p assigned to the variables. If length 2, the proportions of the inertia assigned to the cases and to the variables, respectively.

Details

The inertia of a singular value decomposition \(X=UDV'\) consists in the squares of the singular values (the diagonal elements of \(D\)), and represents the variance, likened to the physical inertia, in the directions of the orthogonal singular vectors (the columns of \(U\) or of \(V\)). Biplots superimpose the projections of the rows and the columns of \(X\) onto these coordinate vectors, scaled by some proportion of the total inertia: \(UD^p\) and \(VD^q\). A biplot is balanced if \(p+q=1\). Read Orlov (2013) for more on conferring inertia in PCA.

recover_conference(), like the other accessors, is an S3 method that is exported for convenience but not intended to be used directly.

Note: In case the "inertia" attribute is a rectangular matrix, one may only be able to confer it entirely to the cases (p = 1) or entirely to the variables (p = 0).

References

Orlov K (2013) Answer to "Algebra of LDA. Fisher discrimination power of a variable and Linear Discriminant Analysis". CrossValidated, accessed 2019-07-26. https://stats.stackexchange.com/a/83114/68743

See also

Examples

# illustrative ordination: correspendence analysis of hair & eye data
haireye_ca <- ordinate(
  as.data.frame(rowSums(HairEyeColor, dims = 2L)),
  cols = everything(), model = ca::ca
)
print(haireye_ca)
#> # A tbl_ord of class 'ca': (4 x 3) x (4 x 3)'
#> # 3 coordinates: Dim1, Dim2, Dim3
#> # 
#> # Rows (standard): [ 4 x 3 | 4 ]
#>     Dim1   Dim2   Dim3 |   .name .mass .dist .inertia
#>                        |   <chr> <dbl> <dbl>    <dbl>
#> 1 -1.10   1.44  -1.09  | 1 Black 0.182 0.551   0.0554
#> 2 -0.324 -0.219  0.957 | 2 Brown 0.483 0.159   0.0123
#> 3 -0.283 -2.14  -1.63  | 3 Red   0.120 0.355   0.0151
#> 4  1.83   0.467 -0.318 | 4 Blond 0.215 0.838   0.151 
#> # 
#> # Columns (standard): [ 4 x 3 | 4 ]
#>     Dim1   Dim2    Dim3 |   .name .mass .dist .inertia
#>                         |   <chr> <dbl> <dbl>    <dbl>
#> 1 -1.08   0.592 -0.424  | 1 Brown 0.372 0.500   0.0931
#> 2  1.20   0.556  0.0924 | 2 Blue  0.363 0.554   0.111 
#> 3 -0.465 -1.12   1.97   | 3 Hazel 0.157 0.289   0.0131
#> 4  0.354 -2.27  -1.72   | 4 Green 0.108 0.386   0.0161
# check distribution of inertia
get_conference(haireye_ca)
#> [1] 0 0
# confer inertia to rows, then to columns
confer_inertia(haireye_ca, "rows")
#> # A tbl_ord of class 'ca': (4 x 3) x (4 x 3)'
#> # 3 coordinates: Dim1, Dim2, Dim3
#> # 
#> # Rows (principal): [ 4 x 3 | 4 ]
#>     Dim1    Dim2    Dim3 |   .name .mass .dist .inertia
#>                          |   <chr> <dbl> <dbl>    <dbl>
#> 1 -0.505  0.215  -0.0555 | 1 Black 0.182 0.551   0.0554
#> 2 -0.148 -0.0327  0.0488 | 2 Brown 0.483 0.159   0.0123
#> 3 -0.130 -0.320  -0.0832 | 3 Red   0.120 0.355   0.0151
#> 4  0.835  0.0696 -0.0162 | 4 Blond 0.215 0.838   0.151 
#> # 
#> # Columns (standard): [ 4 x 3 | 4 ]
#>     Dim1   Dim2    Dim3 |   .name .mass .dist .inertia
#>                         |   <chr> <dbl> <dbl>    <dbl>
#> 1 -1.08   0.592 -0.424  | 1 Brown 0.372 0.500   0.0931
#> 2  1.20   0.556  0.0924 | 2 Blue  0.363 0.554   0.111 
#> 3 -0.465 -1.12   1.97   | 3 Hazel 0.157 0.289   0.0131
#> 4  0.354 -2.27  -1.72   | 4 Green 0.108 0.386   0.0161
confer_inertia(haireye_ca, "columns")
#> # A tbl_ord of class 'ca': (4 x 3) x (4 x 3)'
#> # 3 coordinates: Dim1, Dim2, Dim3
#> # 
#> # Rows (standard): [ 4 x 3 | 4 ]
#>     Dim1   Dim2   Dim3 |   .name .mass .dist .inertia
#>                        |   <chr> <dbl> <dbl>    <dbl>
#> 1 -1.10   1.44  -1.09  | 1 Black 0.182 0.551   0.0554
#> 2 -0.324 -0.219  0.957 | 2 Brown 0.483 0.159   0.0123
#> 3 -0.283 -2.14  -1.63  | 3 Red   0.120 0.355   0.0151
#> 4  1.83   0.467 -0.318 | 4 Blond 0.215 0.838   0.151 
#> # 
#> # Columns (principal): [ 4 x 3 | 4 ]
#>     Dim1    Dim2     Dim3 |   .name .mass .dist .inertia
#>                           |   <chr> <dbl> <dbl>    <dbl>
#> 1 -0.492  0.0883 -0.0216  | 1 Brown 0.372 0.500   0.0931
#> 2  0.547  0.0830  0.00471 | 2 Blue  0.363 0.554   0.111 
#> 3 -0.213 -0.167   0.101   | 3 Hazel 0.157 0.289   0.0131
#> 4  0.162 -0.339  -0.0876  | 4 Green 0.108 0.386   0.0161
# confer inertia symmetrically
(haireye_ca <- confer_inertia(haireye_ca, "symmetric"))
#> # A tbl_ord of class 'ca': (4 x 3) x (4 x 3)'
#> # 3 coordinates: Dim1, Dim2, Dim3
#> # 
#> # Rows (symmetric): [ 4 x 3 | 4 ]
#>     Dim1    Dim2    Dim3 |   .name .mass .dist .inertia
#>                          |   <chr> <dbl> <dbl>    <dbl>
#> 1 -0.746  0.556  -0.246  | 1 Black 0.182 0.551   0.0554
#> 2 -0.219 -0.0846  0.216  | 2 Brown 0.483 0.159   0.0123
#> 3 -0.192 -0.828  -0.368  | 3 Red   0.120 0.355   0.0151
#> 4  1.24   0.180  -0.0718 | 4 Blond 0.215 0.838   0.151 
#> # 
#> # Columns (symmetric): [ 4 x 3 | 4 ]
#>     Dim1   Dim2    Dim3 |   .name .mass .dist .inertia
#>                         |   <chr> <dbl> <dbl>    <dbl>
#> 1 -0.728  0.229 -0.0957 | 1 Brown 0.372 0.500   0.0931
#> 2  0.810  0.215  0.0209 | 2 Blue  0.363 0.554   0.111 
#> 3 -0.315 -0.434  0.445  | 3 Hazel 0.157 0.289   0.0131
#> 4  0.239 -0.878 -0.388  | 4 Green 0.108 0.386   0.0161
# check redistributed inertia
get_conference(haireye_ca)
#> [1] 0.5 0.5
# restore default distribution of inertia
revert_conference(haireye_ca)
#> # A tbl_ord of class 'ca': (4 x 3) x (4 x 3)'
#> # 3 coordinates: Dim1, Dim2, Dim3
#> # 
#> # Rows (standard): [ 4 x 3 | 4 ]
#>     Dim1   Dim2   Dim3 |   .name .mass .dist .inertia
#>                        |   <chr> <dbl> <dbl>    <dbl>
#> 1 -1.10   1.44  -1.09  | 1 Black 0.182 0.551   0.0554
#> 2 -0.324 -0.219  0.957 | 2 Brown 0.483 0.159   0.0123
#> 3 -0.283 -2.14  -1.63  | 3 Red   0.120 0.355   0.0151
#> 4  1.83   0.467 -0.318 | 4 Blond 0.215 0.838   0.151 
#> # 
#> # Columns (standard): [ 4 x 3 | 4 ]
#>     Dim1   Dim2    Dim3 |   .name .mass .dist .inertia
#>                         |   <chr> <dbl> <dbl>    <dbl>
#> 1 -1.08   0.592 -0.424  | 1 Brown 0.372 0.500   0.0931
#> 2  1.20   0.556  0.0924 | 2 Blue  0.363 0.554   0.111 
#> 3 -0.465 -1.12   1.97   | 3 Hazel 0.157 0.289   0.0131
#> 4  0.354 -2.27  -1.72   | 4 Green 0.108 0.386   0.0161