Format a tbl_ord for printing

These methods of base::format() and base::print() render a (usually more) tidy readout of a tbl_ord that is consistent across all original ordination classes.

# S3 method for tbl_ord
format(x, ..., n = NULL, width = NULL, n_extra = NULL)

# S3 method for tbl_ord
print(x, ..., n = NULL, width = NULL, n_extra = NULL)

Arguments

x
A tbl_ord object.

...
Additional arguments.

n
Number of rows to show. If NULL, the default, will print all rows if less than option tibble.print_max. Otherwise, will print tibble.print_min rows.

width
Width of text output to generate. This defaults to NULL, which means use getOption("tibble.width") or (if also NULL) getOption("width"); the latter displays only the columns that fit on one screen. You can also set options(tibble.width = Inf) to override this default and always print all columns, this may be slow for very wide tibbles.

n_extra
Number of extra columns to print abbreviated information for, if the width is too small for the entire tibble. If NULL, the default, will print information about at most tibble.max_extra_cols extra columns.

Details

The format and print methods for class 'tbl_ord' are adapted from those for class 'tbl_df' and for class 'tbl_graph'.

Note: The format() function is tedius but cannot be easily modularized without invoking accessors, annotation, and augmentation multiple times, thereby significantly reducing performance.