Fit an ordination model to columns of a data frame

This is a convenience function to fit an ordination model to a specified subset of columns of a data frame and augment the tbl_ord object with both its intrinsic diagnostics (via [augment_ord()]) and any additional columns of the data frame (via [bind_cols_rows()]).

ordinate(data, cols, model, augment, ...)

Arguments

data
A data frame.

cols
<tidy-select> Columns of data to pass to model.

model
An ordination function whose output is coercible to class 'tbl_ord'. Alternatively, a formula ~ fun(., ...) where fun is such a function and other arguments are explicit, which will be evaluated with data in place of ..

augment
<tidy-select> Columns of data to augment to the row data of the ordination. If missing, all columns not included in cols will be augmented.

...
Additional arguments passed to model.

Examples

# PCA of iris data
ordinate(iris, -Species, princomp, augment = c(Sepal.Width, Species))

#> # A tbl_ord of class 'princomp': (150 x 4) x (4 x 4)'
#> # 4 coordinates: Comp.1, Comp.2, ..., Comp.4
#> # 
#> # Rows (principal): [ 150 x 4 | 2 ]
#>   Comp.1 Comp.2  Comp.3 ... |   Sepal.Width Species
#>                             |         <dbl> <fct>  
#> 1  -2.68  0.319  0.0279     | 1         3.5 setosa 
#> 2  -2.71 -0.177  0.210  ... | 2         3   setosa 
#> 3  -2.89 -0.145 -0.0179     | 3         3.2 setosa 
#> 4  -2.75 -0.318 -0.0316     | 4         3.1 setosa 
#> 5  -2.73  0.327 -0.0901     | 5         3.6 setosa 
#> # … with 145 more rows
#> # 
#> # Columns (standard): [ 4 x 4 | 3 ]
#>    Comp.1  Comp.2  Comp.3 ... |   .name        .center .scale
#>                               |   <chr>          <dbl>  <dbl>
#> 1  0.361   0.657   0.582      | 1 Sepal.Length    5.84      1
#> 2 -0.0845  0.730  -0.598  ... | 2 Sepal.Width     3.06      1
#> 3  0.857  -0.173  -0.0762     | 3 Petal.Length    3.76      1
#> 4  0.358  -0.0755 -0.546      | 4 Petal.Width     1.20      1

ordinate(iris, 1:4, ~ prcomp(., center = TRUE, scale. = TRUE))

#> # A tbl_ord of class 'prcomp': (150 x 4) x (4 x 4)'
#> # 4 coordinates: PC1, PC2, ..., PC4
#> # 
#> # Rows (principal): [ 150 x 4 | 1 ]
#>     PC1    PC2     PC3 ... |   Species
#>                            |   <fct>  
#> 1 -2.26 -0.478  0.127      | 1 setosa 
#> 2 -2.07  0.672  0.234  ... | 2 setosa 
#> 3 -2.36  0.341 -0.0441     | 3 setosa 
#> 4 -2.29  0.595 -0.0910     | 4 setosa 
#> 5 -2.38 -0.645 -0.0157     | 5 setosa 
#> # … with 145 more rows
#> # 
#> # Columns (standard): [ 4 x 4 | 3 ]
#>      PC1     PC2    PC3 ... |   .name        .center .scale
#>                             |   <chr>          <dbl>  <dbl>
#> 1  0.521 -0.377   0.720     | 1 Sepal.Length    5.84  0.828
#> 2 -0.269 -0.923  -0.244 ... | 2 Sepal.Width     3.06  0.436
#> 3  0.580 -0.0245 -0.142     | 3 Petal.Length    3.76  1.77 
#> 4  0.565 -0.0669 -0.634     | 4 Petal.Width     1.20  0.762

# LRA of arrest data
ordinate(USArrests, cols = c(Murder, Rape, Assault), lra)

#> # A tbl_ord of class 'lra': (50 x 2) x (3 x 2)'
#> # 2 coordinates: LRSV1 and LRSV2
#> # 
#> # Rows (standard): [ 50 x 2 | 2 ]
#>    LRSV1  LRSV2 |   .name      UrbanPop
#>                 |   <chr>         <int>
#> 1 -0.680  0.930 | 1 Alabama          58
#> 2  0.930 -0.625 | 2 Alaska           48
#> 3 -0.330 -1.31  | 3 Arizona          80
#> 4 -0.351  0.277 | 4 Arkansas         50
#> 5  0.552 -1.00  | 5 California       91
#> # … with 45 more rows
#> # 
#> # Columns (standard): [ 3 x 2 | 1 ]
#>    LRSV1  LRSV2 |   .name  
#>                 |   <chr>  
#> 1  0.283  4.96  | 1 Murder 
#> 2  2.88  -0.366 | 2 Rape   
#> 3 -0.371 -0.181 | 3 Assault

# CA of hair & eye color data
haireye <- as.data.frame(rowSums(HairEyeColor, dims = 2L))
ordinate(haireye, cols = everything(), model = ca::ca, augment = everything())

#> # A tbl_ord of class 'ca': (4 x 3) x (4 x 3)'
#> # 3 coordinates: Dim1, Dim2, Dim3
#> # 
#> # Rows (standard): [ 4 x 3 | 8 ]
#>     Dim1   Dim2   Dim3 |   .name .mass .dist .inertia Brown
#>                        |   <chr> <dbl> <dbl>    <dbl> <dbl>
#> 1 -1.10   1.44  -1.09  | 1 Black 0.182 0.551   0.0554    68
#> 2 -0.324 -0.219  0.957 | 2 Brown 0.483 0.159   0.0123   119
#> 3 -0.283 -2.14  -1.63  | 3 Red   0.120 0.355   0.0151    26
#> 4  1.83   0.467 -0.318 | 4 Blond 0.215 0.838   0.151      7
#> # … with 3 more variables:
#> #   Blue <dbl>, Hazel <dbl>,
#> #   Green <dbl>
#> # 
#> # Columns (standard): [ 4 x 3 | 4 ]
#>     Dim1   Dim2    Dim3 |   .name .mass .dist .inertia
#>                         |   <chr> <dbl> <dbl>    <dbl>
#> 1 -1.08   0.592 -0.424  | 1 Brown 0.372 0.500   0.0931
#> 2  1.20   0.556  0.0924 | 2 Blue  0.363 0.554   0.111 
#> 3 -0.465 -1.12   1.97   | 3 Hazel 0.157 0.289   0.0131
#> 4  0.354 -2.27  -1.72   | 4 Green 0.108 0.386   0.0161

# lSVD of finches data
data(finches, package = "cooccur")
ordinate(finches, everything(), logisticSVD_ord)

#> # A tbl_ord of class 'lsvd': (13 x 2) x (17 x 2)'
#> # 2 coordinates: LSC1 and LSC2
#> # 
#> # Rows: [ 13 x 2 | 1 ]
#>    LSC1  LSC2 |   .name                
#>               |   <chr>                
#> 1 27.4  -31.3 | 1 Geospiza magnirostris
#> 2 47.5   20.3 | 2 Geospiza fortis      
#> 3 49.2  -15.9 | 3 Geospiza fuliginosa  
#> 4 -6.31 -32.6 | 4 Geospiza difficilis  
#> 5 57.6   67.7 | 5 Geospiza scandens    
#> # … with 8 more rows
#> # 
#> # Columns: [ 17 x 2 | 2 ]
#>    LSC1    LSC2 |   .name        .mu
#>                 |   <chr>      <dbl>
#> 1 0.290  0.0501 | 1 Seymour    -9.73
#> 2 0.290  0.0501 | 2 Baltra     -9.73
#> 3 0.149 -0.177  | 3 Isabella   11.4 
#> 4 0.188 -0.398  | 4 Fernandina 13.4 
#> 5 0.216 -0.0734 | 5 Santiago    7.22
#> # … with 12 more rows