Wrappers for lossy ordination methods

These *_ord functions wrap functions imported from other packages with modifications for use with 'tbl_ord' methods. Some parameters are hidden from the user and set to settings required for these methods, some matrix outputs are given row or column names to be used by them, and new '*_ord' S3 class attributes are added to enable them.

eigen_ord(x, symmetric = isSymmetric.matrix(x))

svd_ord(x, nu = min(dim(x)), nv = min(dim(x)))

cmdscale_ord(d, k = 2, add = FALSE)

logisticSVD_ord(
  x,
  k = 2,
  quiet = TRUE,
  max_iters = 1000,
  conv_criteria = 1e-05,
  random_start = FALSE,
  ...,
  partial_decomp = TRUE,
  main_effects = TRUE
)

logisticPCA_ord(
  x,
  k = 2,
  m = 4,
  quiet = TRUE,
  partial_decomp = FALSE,
  max_iters = 1000,
  conv_criteria = 1e-05,
  random_start = FALSE,
  ...,
  main_effects = TRUE
)

convexLogisticPCA_ord(
  x,
  k = 2,
  m = 4,
  quiet = TRUE,
  partial_decomp = FALSE,
  max_iters = 1000,
  conv_criteria = 1e-06,
  random_start = FALSE,
  ...,
  main_effects = TRUE,
  ss_factor = 4
)

Arguments

x
a numeric or complex matrix whose spectral decomposition is to be computed. Logical matrices are coerced to numeric.

symmetric
if TRUE, the matrix is assumed to be symmetric (or Hermitian if complex) and only its lower triangle (diagonal included) is used. If symmetric is not specified, isSymmetric(x) is used.

nu
the number of left singular vectors to be computed. This must between 0 and n = nrow(x).

nv
the number of right singular vectors to be computed. This must be between 0 and p = ncol(x).

d
a distance structure such as that returned by dist or a full symmetric matrix containing the dissimilarities.

k
the maximum dimension of the space which the data are to be represented in; must be in \(\{1, 2, \ldots, n-1\}\).

add
logical indicating if an additive constant \(c*\) should be computed, and added to the non-diagonal dissimilarities such that the modified dissimilarities are Euclidean.

quiet
logical; whether the calculation should give feedback

max_iters
number of maximum iterations

conv_criteria
convergence criteria. The difference between average deviance in successive iterations

random_start
logical; whether to randomly inititalize the parameters. If FALSE, function will use an eigen-decomposition as starting value

...
Additional parameters passed to original functions.

partial_decomp
logical; if TRUE, the function uses the rARPACK package to more quickly calculate the eigen-decomposition. This is usually faster than standard eigen-decomponsition when ncol(x) > 100 and k is small

main_effects
logical; whether to include main effects in the model

m
value to approximate the saturated model. If m = 0, m is solved for

ss_factor
step size multiplier. Amount by which to multiply the step size. Quadratic convergence rate can be proven for ss_factor = 1, but I have found higher values sometimes work better. The default is ss_factor = 4. If it is not converging, try ss_factor = 1.

Details

The following table summarizes the wrapped functions:

Original function	Hide params	Add names	New class

base::eigen()	Yes	Yes	Yes

base::svd()	Yes	Yes	Yes

stats::cmdscale()	Yes	No	Yes

logisticPCA::logisticSVD()	No	Yes	No

logisticPCA::logisticPCA()	No	Yes	No

logisticPCA::convexLogisticPCA()	No	Yes	No